Bug fix in QwHelicity (well, kind of...).  I don't know why this bit me only now, but there was a for loop until size_t(fPatternNumber-fPatternNumberOld), which took a really long time for some incorrectly read-out events (lost helicity gates in the Compton DAQ).
git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2187 b9700a98-9937-0410-b777-f318d6eb5bff

+++ Parity/src/QwHelicity.cc
@@ -1406,7 +1406,7 @@ void QwHelicity::RunPredictor()
     */
 
 
-    for( size_t i=0;i<size_t(fPatternNumber- fPatternNumberOld);i++) //got a new pattern
+    for (int i = 0; i < fPatternNumber - fPatternNumberOld; i++) //got a new pattern
       {
 	fPreviousPatternPolarity = fActualPatternPolarity;
 	fActualPatternPolarity   = GetRandbit(iseed_Actual);
